refactor(PR3-c10a): move exampleparent RenderChildren to snapshot/; delete state-package mirror
